Fix selecting arrows
Even though I was drawing highlighted arrows on top of other arrows, it seems that once an arrow receives a single click, all arrows go back to being completely unhighlighted and we revert back to drawing arrows on top of each other in an arbitrary manner.

This caused a problem with double-clicking highlighted arrows because the first click will cause the arrow to no longer be highlighted and the second click might land on a different arrow because the original arrow of interest has been buried now that it is no longer highlighted.

I added a category to highlight status, "IS_SELECTED" and now a selected arrow is always drawn last, making it sit on top of all other arrows.  So the first click now raises an arrow to the top of the stack, making it a prime target for that second click.