Fixed bug loading old specs caused by changing variables names
We just ran headlong into issues with serialization in Java.

We've been naively renaming a few fields in some of the classes and discovered this morning that Java does not properly deserialize a persisted objects once field names are changed.

What we're trying to say is that several of our recent pull requests broke the ability to read in the ETL spec document, but only this morning did the process actually throw an exception and brought it to our attention.

This pull request changes all the field names back to what they were originally and should reinstate the ability to load in an ETL spec file.