Don't allow dragging when filtering
Sidesteps an odd issue where once we filter the source or target, if we click on a source or target box, the box is considered "dragged" and we reorder all the items based on that drag, but not all items are visible and so the sort order of all the boxes becomes corrupted.

Our solution is to just not allow people to reorder (i.e. drag) boxes when the filter is active.